+++ io.h
@@ -16,6 +16,6 @@ public:
     void saveState(System *system);
     void close();
 
-    void writeStatistics(float time, float kineticEnergy, float potentialEnergy, float pressure, float temperature);
+    void writeStatistics(float time, double kineticEnergy, double potentialEnergy, double pressure, double temperature);
     void writePerformance(unsigned int timestep);
 };
